refactor: normalize amount and available_amount to int across architectures (#27)
* refactor: normalize amount and available_amount to int across architectures

* feat: add numconv utility for int-to-int32 conversion and integrate it across services for safer type handling

* refactor(item-service): extract item aggregation logic into reusable function

* docs(database-schema): update `available_amount` type from BIGINT to INT

* fix(transaction-service): handle invalid input errors in gRPC item client

* fix(transaction-service): handle proto mapping errors and add overflow test cases

* test(api-gateway): add conditional test case for `available_amount` overflow on 64-bit systems

* test(api-gateway): add 64-bit specific test case for `amount` overflow in transaction client

+++ microservices/transaction-service/internal/domain/transaction.go
@@ -12,5 +12,5 @@ type Transaction struct {
 
 type TransactionItem struct {
 	ItemID string
-	Amount int64
+	Amount int
 }
